+++ src/main.js
@@ -5,6 +5,9 @@ import store from './store'
 import '@/style/common.less'
 import '@/utils/vant-ui'
 
+import Modal from '@/utils/modal'
+Vue.use(Modal)
+
 Vue.config.productionTip = false
 
 new Vue({